eu logo input oma komponentti
